% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superheat.R
\name{superheat}
\alias{superheat}
\title{Generate supervised heatmaps.}
\usage{
superheat(X, yt = NULL, yr = NULL, yt.plot.type = c("scatter", "bar",
  "boxplot", "scattersmooth", "smooth", "scatterline", "line"),
  yr.plot.type = c("scatter", "bar", "boxplot", "scattersmooth", "smooth",
  "scatterline", "line"), smooth.heat = FALSE, scale = FALSE,
  membership.rows = NULL, membership.cols = NULL, n.clusters.rows = NULL,
  n.clusters.cols = NULL, cluster.rows = FALSE, cluster.cols = FALSE,
  clustering.method = c("kmeans", "hierarchical"), cluster.box = TRUE,
  box.size = 0.5, box.col = "black", smoothing.method = c("loess", "lm"),
  smooth.se = TRUE, legend = TRUE, order.cols = NULL, order.rows = NULL,
  left.heat.label = NULL, bottom.heat.label = NULL, yt.axis = T,
  yr.axis = T, yt.num.ticks = 3, yr.num.ticks = 3, yt.plot.size = 0.3,
  yr.plot.size = 0.3, yt.axis.name = NULL, yr.axis.name = NULL,
  yr.axis.size = 10, yt.axis.size = 10, yr.axis.name.size = 10,
  yt.axis.name.size = 10, bottom.text.size = 5, left.text.size = 5,
  bottom.text.angle = NULL, left.text.angle = NULL,
  bottom.label.size = 0.1, left.label.size = 0.1,
  heat.col.scheme = c("red", "purple", "blue", "grey", "green"),
  heat.pal = NULL, heat.pal.values = NULL, left.label.pal = NULL,
  bottom.label.pal = NULL, left.label.text.col = NULL,
  bottom.label.text.col = NULL, yt.obs.col = NULL, yr.obs.col = NULL,
  yt.pal = NULL, yr.pal = NULL, yt.bar.col = NULL, yr.bar.col = NULL,
  yt.point.size = 2, yt.point.alpha = 1, yr.point.size = 2,
  yr.point.alpha = 1, yr.line.size = NULL, yt.line.size = NULL,
  legend.size = 2, padding = 1, title = NULL, title.size = 5,
  print.plot = TRUE)
}
\arguments{
\item{X}{an optional matrix or data frame containing the variables in the
columns and the observations in the rows.}

\item{yt}{an optional argument (list or vector) specifying the plot above
the heatmap. If a vector is given it must be of length
\code{ncol(X)} and the order of the values in \code{yt} should
correspond to the order of values provided in the columns of \code{X}.
If a list is given, each element of the list must correspond to a
cluster or variable in the plot.}

\item{yr}{an optional argument (list or vector) specifying the plot to the
right of the heatmap. If a vector is given it must be of length
\code{nrow(X)} and the order of the values in \code{yr} should
correspond to the order of values provided in the rows of \code{X}.
If a list is given, each element of the list must correspond to a
cluster or variable in the plot.}

\item{yt.plot.type}{character specifying the top plot type. The default is
"scatter", and other options include "bar", "scattersmooth",
"smooth", "boxplot", "scatterline" and "line".}

\item{yr.plot.type}{character specifying the right plot type. The default is
"scatter", and other options include "bar", "scattersmooth",
"smooth", "boxplot", "scatterline" and "line".}

\item{smooth.heat}{logical. If \code{TRUE}, then the color of the heatmap is
smoothed within clusters.}

\item{scale}{logical. If \code{TRUE}, the columns of X are centered and scaled.}

\item{membership.rows}{an optional vector specifying the cluster membership
of the rows/observations in X. If not specified, the default settings
are to perform hierarchical clustering on the rows.}

\item{membership.cols}{an optional vector specifying the cluster membership
of the columns/variables in X. If not specified, the default settings
are to present the columns in the order they are provided in \code{X}.}

\item{n.clusters.rows}{If \code{membership.rows} is not specified and
\code{cluster.rows = TRUE} (the default), then this argument must be
used to specify how many row clusters to generate.}

\item{n.clusters.cols}{If \code{membership.cols} is not specified and
\code{cluster.cols = TRUE}, then this argument must be used to
specify how many column clusters to generate.}

\item{cluster.rows}{should clustering to be performed on the rows? If
\code{FALSE}, the rows are presented in the same order that they
appear in \code{X} or in the order specified by \code{order.rows}.}

\item{cluster.cols}{should clustering to be performed on the columns? If
\code{FALSE}, the columns are presented in the same order that they
appear in \code{X} or in the order specified by \code{order.cols}.}

\item{clustering.method}{the clustering method to use. The default ("kmeans")
is to use K-means clustering, the alternative option ("hierarchical")
performs heirarchical clustering. Antoher alternative is to provide
a membership vector.}

\item{cluster.box}{logical. If \code{TRUE} (default), then boxes are plotted
around the clusters in the heatmap.}

\item{box.size}{the thickness of the cluster.box lines. The default is 0.5.}

\item{box.col}{the colour of the cluster.box lines.}

\item{smoothing.method}{if \code{plot.type = "scattersmooth"} or \code{"smooth"}
this argument specifies the smoothing method to use. The default is
\code{loess}. The alternative option is \code{lm}.}

\item{smooth.se}{logical. Should the smoothed curve have a standard error
around it?}

\item{legend}{logical. If set to \code{FALSE}, then no legend is given.}

\item{order.cols}{a vector of specifying the ordering of the columns of
\code{X}. If clustering is performed (or a membership vector is
provided), then this vector specifies the order within the clusters.
Note that this vector must be a rearranged 1:ncol(X) vector.}

\item{order.rows}{a vector of specifying the ordering of the rows of
\code{X}. If clustering is performed (or a membership vector is
provided), then this vector specifies the order within the clusters.
Note that this vector must be a rearranged 1:nrow(X) vector.}

\item{left.heat.label}{the type of label provided to the left of the heatmap.
The default is "cluster" (which provides the cluster names) if
clustering was performed on the rowss. Otherwise, the default is
"variable" (which provides the variable names). The final option
("none") removes the label all together.}

\item{bottom.heat.label}{the type of label provided to the left of the heatmap.
The default is "cluster" (which provides the cluster names) if
clustering was performed on the columns. Otherwise, the default is
"variable" (which provides the variable names). The final option
("none") removes the label all together.}

\item{yt.axis}{logical. If \code{TRUE} (default), and there is a
scatterplot plotted above the heatmap (\code{yt} is provided), then
the plot will have an axis.}

\item{yr.axis}{logical. If \code{TRUE} (default), and there is a
scatterplot plotted to the right of the heatmap (\code{yr} is
provided), then the plot will have an axis.}

\item{yt.num.ticks}{the number of ticks on the \code{yt} axis. This does
not always work perfectly.}

\item{yr.num.ticks}{the number of ticks on the \code{yr} axis. This does
not always work perfectly.}

\item{yt.plot.size}{a number specifying the size of the scatter plot on top
of the heatmap.}

\item{yr.plot.size}{a number specifying the size of the scatter plot to the
right of the heatmap.}

\item{yt.axis.name}{a character specifying the axis label.}

\item{yr.axis.name}{a character specifying the axis label.}

\item{yr.axis.size}{a number specifying the size of the numbers on
the axis.}

\item{yt.axis.size}{a number specifying the size of the numbers on
the axis.}

\item{yr.axis.name.size}{a number specifying the size of the axis name.}

\item{yt.axis.name.size}{a number specifying the size of the axis name.}

\item{bottom.text.size}{the size of the bottom heatmap label text. The
default is 5.}

\item{left.text.size}{the size of the left heatmap label text. The
default is 5.}

\item{bottom.text.angle}{number of degrees to rotate the text on the bottom
cluster/variable labels. The default is 0.}

\item{left.text.angle}{number of degrees to rotate the text on the left
cluster/variable labels. The default is 90.}

\item{bottom.label.size}{a number specifying the size of the bottom
cluster/variable labels.}

\item{left.label.size}{a number specifying the size of the left
cluster/variable labels.}

\item{heat.col.scheme}{A character specifying the heatmap colour scheme.
The default is "red", and other options include "purple", "blue",
"grey" and "green". If you wish to supply your own colour scheme,
you can use the \code{heat.pal} argument.}

\item{heat.pal}{a vector specifying a manual heatmap color palette. This
corresponds to the \code{color} argumnet for the ggplot2
\code{\link[ggplot2]{scale_color_gradientn}} function.}

\item{heat.pal.values}{a vector specifying the location of each color in the
color palatte. Each entry should be a number between 0 and 1. This
corresponds to the \code{values} argumnet for the ggplot2
\code{\link[ggplot2]{scale_color_gradientn}} function. The default
values correspond to the equally spaced quantiles.}

\item{left.label.pal}{a vector specifying the left cluster/variable label color
palette.}

\item{bottom.label.pal}{a vector specifying the bottom cluster/variable label color
palette.}

\item{left.label.text.col}{a character or character vector specifying the
left cluster/variable label text color.}

\item{bottom.label.text.col}{a character or character vector specifying the
bottom cluster/variable label text color.}

\item{yt.obs.col}{a vector specifying the colour of individual points in the
top scatterplot.}

\item{yr.obs.col}{a vector specifying the colour of individual points in the
right scatterplot.}

\item{yt.pal}{a vector the same length as the number of clusters which specifies
the colour of each cluster in \code{yt}.}

\item{yr.pal}{a vector the same length as the number of clusters which specifies
the colour of each cluster in \code{yr}.}

\item{yt.bar.col}{a character which specifies the colour of the boundary of the
bars in the barplot of \code{yt}.}

\item{yr.bar.col}{a character which specifies the colour of the boundary of the
bars in the barplot of \code{yr}.}

\item{yt.point.size}{the size of the points in the top scatterplot. The
default is 2.}

\item{yt.point.alpha}{the transparancy of the points in the top scatterplot.
The default is 1, which corresponds to no transparancy.}

\item{yr.point.size}{the size of the points in the right scatterplot. The
default is 2.}

\item{yr.point.alpha}{the transparancy of the points in the top scatterplot.
The default is 1, which corresponds to no transparancy.}

\item{yr.line.size}{the thickness of the (smoothing) line in the right plot.}

\item{yt.line.size}{the thickness of the (smoothing) line in the top plot.}

\item{legend.size}{a number specifying the size of the legend. The default
is 2.}

\item{padding}{the amount (in cm) of white space (padding) around the plot.
The default is 1cm.}

\item{title}{an optional character string specifying a main heading.}

\item{title.size}{the size of the title. The default is 5.}

\item{print.plot}{logical. If \code{TRUE} (default), the plot will be
printed to the screen when the function is called.}

\item{line.size}{the thickness of the line if \code{plot.type} is
\code{"scattersmooth"}, \code{"smooth"}, \code{"scatterline"} or
\code{"line"}. The default is 0.5.}
}
\value{
\code{plot} a plot with the properties specified by the above arguments.

\code{membership.cols} the column cluster membership vector

\code{membership.rows} the row cluster membership vector
}
\description{
superheat is used to generate an exploratory plot which consists of a
       clustered matrix X displayed in the form of a heatmap. Scatterplots
       can be plotted above and to the right of the heatmap. This plot allows
       for the user to view X versus y for a multivariate X.
}
\examples{
# plot a heatmap of the numerical iris variables
# cluster by species and plot Sepal.Length on the right
superheat(X = iris[,-c(1, 5)],
            yr = iris[,1],
            yr.axis.name = "Sepal.Length",
            membership.rows = iris$Species)
}

