% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labels.R
\name{generate_cluster_label}
\alias{generate_cluster_label}
\title{Generate clustered heatmap labels.}
\usage{
generate_cluster_label(membership, location = c("bottom", "left"),
  label.pal = NULL, label.text.col = NULL, bottom.text.size = 5,
  left.text.size = 5, text.angle = NULL)
}
\arguments{
\item{membership}{a vector specifying the cluster membership.}

\item{location}{will these labels be on the bottom ("bottom") or on the left
("left)?}

\item{label.pal}{a vector specifying the cluster/variable label color palette.}

\item{label.text.col}{a character or character vector specifying the
cluster/variable label text color.}

\item{bottom.text.size}{the size of the bottom heatmap label text. The
default is 5.}

\item{left.text.size}{the size of the left heatmap label text. The
default is 5.}

\item{text.angle}{number of degrees to rotate the text on the left
cluster/variable labels.}
}
\value{
A ggplot2 object of heatmap labels.
}
\description{
Generate clustered heatmap labels.
}

