% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{generate_var_label}
\alias{generate_var_label}
\title{Generate heatmap variable labels.}
\usage{
generate_var_label(names, location = c("bottom", "left"), label.col = NULL,
  label.text.col = NULL, bottom.label.text.size = 5,
  left.label.text.size = 5, text.angle = NULL)
}
\arguments{
\item{names}{a vector specifying the label names.}

\item{location}{will these labels be on the bottom ("bottom") or on the left
("left)?}

\item{label.col}{a vector specifying the cluster/variable label color palette.}

\item{label.text.col}{a character or character vector specifying the
cluster/variable label text color.}

\item{bottom.label.text.size}{the size of the bottom heatmap label text. The
default is 5.}

\item{left.label.text.size}{the size of the left heatmap label text. The
default is 5.}

\item{text.angle}{number of degrees to rotate the text on the left
cluster/variable labels.}
}
\value{
A ggplot2 object of heatmap labels.
}
\description{
Generate heatmap variable labels.
}

